Refactored DTO packages
