+++ src/main/java/com/it_academyproject/controllers/DTOs/statsDTOs/DTOStudentsPerItinerary.java
@@ -1,4 +1,4 @@
-package com.it_academyproject.controllers.statsDTOs;
+package com.it_academyproject.controllers.DTOs.statsDTOs;
 
 public class DTOStudentsPerItinerary {
